requirements.txt
treamlit
tensorflow
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib